Add structure Add utils data handling tests
